requirements.txt
matplotlib~=3.5.1
numpy~=1.22.0
Pillow~=9.0.0
mplsoccer
highlight-text